+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Session Metadata
 
+## 2.4.0.0 - 2024-03-15
+* Can now write metadata for flats, darks, and bias frames (off by default)
+* Output date format for exposure start times is now yyyy-MM-dd HH:mm
+
+## 2.3.4.0 - 2023-10-27
+* Added FWHM and eccentricity (calculated by Hocus Focus) to the output metrics.
+
+## 2.3.0.0 - 2022-12-22
+* Migrated to .NET 7 for NINA 3
+
 ## 1.3.0.0 - 2002-08-XX
 * Added much better support for specifying the location of metadata files
 * Added output of weather metrics (separate file)